Version 0.5.0-pre
Probing compares each measured height against the heights already measured
around it. Elapsed time cannot judge a probe: a long traverse to the next
point takes as long as a bad reading. A height past
ProbeHeightDeviationToleranceMm is re-probed rather than recorded. The tool
lifts to safe height before the run pauses, and the job stops if that lift is
not confirmed.

Closing the serial port does not stop GRBL; it keeps running the moves in its
planner buffer. Every path that ends a run reaches
ProbeController.CleanupAsync, which stops the machine and then lifts the tool
to safe height. CleanupAsync reports an error when the lift is not confirmed.

AppState.CurrentProbeGrid is the one answer to "is there probe data": the
grid in memory, or the autosave when nothing is loaded.
AppState.ReadUsableAutosave is the only code that reads the autosave file.
The mill preflight and the probe screen read the same answer, so a complete
map sitting unapplied in the autosave refuses the job rather than letting it
cut with no height correction. ProbeGrid.Load applies the same checks as the
ProbeGrid constructor, so a file with a non-finite origin is refused at load
rather than matching every job. Probing requires a work zero, and the
toolpath cannot change while a run is in progress.

In the web layer: an answer to a prompt names the prompt it answers, and a
start endpoint reports whether it started. The browser's screen lock and
pause button derive from what the server reports. HTTP and WebSocket commands
share one table, and WebSocket sends are serialized per client. A request
body is capped at 1 MB and an upload at 16 MB. Exception text stays off the
operator's screen.

Milling holds at a prompt until the machine reports the door closed, and
starts only when the preflight checks pass.

Each fact has one home: Machine.OperatingMode names what the machine is
driven to do and Machine.Connected answers whether there is a link, the
terminal reads the controller's state, the grid reports its extremes from its
nodes, and the height gradient is computed once for both views.

Seventeen tests assert conditions that fail when the behavior is wrong, and
the web layer has an end-to-end harness: FakeGrbl, a GRBL stand-in on a
loopback port that the real Machine connects to.

coppercli targets GRBL 1.1f.

+++ coppercli.Core/Controllers/MillingOptions.cs
@@ -9,7 +9,7 @@ namespace coppercli.Core.Controllers
     {
         /// <summary>Builds milling options. The RequireHoming = not-yet-homed rule lives
         /// here so both front ends enforce it identically.</summary>
-        public static MillingOptions Create(string? filePath, float depthAdjustment,
+        public static MillingOptions Create(string? filePath, double depthAdjustment,
             bool machineIsHomed)
         {
             return new MillingOptions
@@ -28,7 +28,7 @@ namespace coppercli.Core.Controllers
         /// Depth adjustment in mm (negative = deeper).
         /// Applied as offset to work coordinate Z origin.
         /// </summary>
-        public float DepthAdjustment { get; set; }
+        public double DepthAdjustment { get; set; }
 
         /// <summary>
         /// Whether to home the machine if not already homed.
